+++ app/code/core/Mage/Eav/Model/Entity/Abstract.php
@@ -1,23 +1,15 @@
 <?php
 
 /**
- * OpenMage
- *
- * This source file is subject to the Open Software License (OSL 3.0)
- * that is bundled with this package in the file LICENSE.txt.
- * It is also available at https://opensource.org/license/osl-3-0-php
- *
- * @category   Mage
+ * @copyright  For copyright and license information, read the COPYING.txt file.
+ * @link       /COPYING.txt
+ * @license    Open Software License (OSL 3.0)
  * @package    Mage_Eav
- * @copyright  Copyright (c) 2006-2020 Magento, Inc. (https://www.magento.com)
- * @copyright  Copyright (c) 2017-2024 The OpenMage Contributors (https://www.openmage.org)
- * @license    https://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
 /**
  * Entity/Attribute/Model - entity abstract
  *
- * @category   Mage
  * @package    Mage_Eav
  */
 abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_Abstract implements Mage_Eav_Model_Entity_Interface
@@ -189,6 +181,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
         if (is_string($this->_read)) {
             $this->_read = Mage::getSingleton('core/resource')->getConnection($this->_read);
         }
+
         return $this->_read;
     }
 
@@ -202,6 +195,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
         if (is_string($this->_write)) {
             $this->_write = Mage::getSingleton('core/resource')->getConnection($this->_write);
         }
+
         return $this->_write;
     }
 
@@ -271,6 +265,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
         if (empty($this->_type)) {
             throw Mage::exception('Mage_Eav', Mage::helper('eav')->__('Entity is not initialized'));
         }
+
         return $this->_type;
     }
 
@@ -341,7 +336,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
      *
      * If attribute is not found false is returned
      *
-     * @param string|int|Mage_Core_Model_Config_Element $attribute
+     * @param string|int|Mage_Core_Model_Config_Element|Mage_Eav_Model_Entity_Attribute_Abstract $attribute
      * @return Mage_Eav_Model_Entity_Attribute_Abstract|Mage_Catalog_Model_Resource_Eav_Attribute|false
      */
     public function getAttribute($attribute)
@@ -352,6 +347,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
             if (isset($this->_attributesById[$attributeId])) {
                 return $this->_attributesById[$attributeId];
             }
+
             $attributeInstance = Mage::getSingleton('eav/config')->getAttribute($this->getEntityType(), $attributeId);
             if ($attributeInstance) {
                 $attributeCode = $attributeInstance->getAttributeCode();
@@ -362,6 +358,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
             if (isset($this->_attributesByCode[$attributeCode])) {
                 return $this->_attributesByCode[$attributeCode];
             }
+
             $attributeInstance = Mage::getSingleton('eav/config')
                 ->getAttribute($this->getEntityType(), $attributeCode);
             if ($attributeInstance && !$attributeInstance->getAttributeCode() && in_array($attribute, $this->getDefaultAttributes())) {
@@ -395,9 +392,10 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
             $attributeId = $attribute->getAttributeId();
         }
 
-        if (!$attribute->getAttributeCode()) {
+        if (isset($attributeCode) && !$attribute->getAttributeCode()) {
             $attribute->setAttributeCode($attributeCode);
         }
+
         if (!$attribute->getAttributeModel()) {
             $attribute->setAttributeModel($this->_getDefaultAttributeModel());
         }
@@ -463,6 +461,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
         if ($flag !== null) {
             $this->_isPartialLoad = (bool) $flag;
         }
+
         return $result;
     }
 
@@ -478,6 +477,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
         if ($flag !== null) {
             $this->_isPartialSave = (bool) $flag;
         }
+
         return $result;
     }
 
@@ -594,6 +594,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
     public function walkAttributes($partMethod, array $args = [])
     {
         $methodArr = explode('/', $partMethod);
+        $part = '';
         switch (count($methodArr)) {
             case 1:
                 $part   = 'attribute';
@@ -605,6 +606,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
                 $method = $methodArr[1];
                 break;
         }
+
         $results = [];
         foreach ($this->getAttributesByCode() as $attrCode => $attribute) {
             if (isset($args[0]) && is_object($args[0]) && !$this->_isApplicableAttribute($args[0], $attribute)) {
@@ -629,7 +631,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
                     break;
             }
 
-            if (!$this->_isCallableAttributeInstance($instance, $method, $args)) {
+            if (!isset($instance, $method) || !$this->_isCallableAttributeInstance($instance, $method, $args)) {
                 continue;
             }
 
@@ -706,6 +708,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
             if (!$table) {
                 $table = Mage_Eav_Model_Entity::DEFAULT_ENTITY_TABLE;
             }
+
             $this->_entityTable = Mage::getSingleton('core/resource')->getTableName($table);
         }
 
@@ -773,6 +776,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
                     $prefix = Mage_Eav_Model_Entity::DEFAULT_ENTITY_TABLE;
                 }
             }
+
             $this->_entityTablePrefix = $prefix;
         }
 
@@ -812,6 +816,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
                 $errors[$attributeCode] = $error;
             }
         }
+
         if (!$errors) {
             return true;
         }
@@ -866,6 +871,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
                 $date  = new Zend_Date($value, Varien_Date::DATE_INTERNAL_FORMAT);
                 $value = $date->toString(Varien_Date::DATETIME_INTERNAL_FORMAT);
             }
+
             $bind = [
                 'entity_type_id' => $this->getTypeId(),
                 'attribute_id'   => $attribute->getId(),
@@ -877,12 +883,14 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
                 ->where('attribute_id = :attribute_id')
                 ->where('value = :value');
         }
+
         $data = $adapter->fetchCol($select, $bind);
 
         if ($object->getId()) {
             if (isset($data[0])) {
                 return $data[0] == $object->getId();
             }
+
             return true;
         }
 
@@ -963,6 +971,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
             $select = $this->_getLoadAttributesSelect($object, $table);
             $selects[$eavType][] = $this->_addLoadAttributesSelectFields($select, $table, $eavType);
         }
+
         /** @var Mage_Eav_Model_Resource_Helper_Mysql4 $helper */
         $helper = Mage::getResourceHelper('eav');
         $selectGroups = $helper->getLoadAttributesSelectGroups($selects);
@@ -1095,7 +1104,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
      */
     protected function _getOrigObject($object)
     {
-        $className  = get_class($object);
+        $className  = $object::class;
         $origObject = new $className();
         $origObject->setData([]);
         $this->load($origObject, $object->getData($this->getEntityIdField()));
@@ -1149,6 +1158,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
 
         $staticFields   = $this->_getReadAdapter()->describeTable($this->getEntityTable());
         $staticFields   = array_keys($staticFields);
+
         $attributeCodes = array_keys($this->_attributesByCode);
 
         foreach ($newData as $k => $v) {
@@ -1158,6 +1168,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
             if (is_numeric($k) || is_array($v)) {
                 continue;
             }
+
             /**
              * Check if data key is presented in static fields or attribute codes
              */
@@ -1323,6 +1334,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
                 $adapter->insert($entityTable, $entityRow);
                 $entityId = $adapter->lastInsertId($entityTable);
             }
+
             $newObject->setId($entityId);
         } else {
             $where = sprintf('%s=%d', $adapter->quoteIdentifier($entityIdField), $entityId);
@@ -1469,6 +1481,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
         if (!isset(self::$_attributeBackendTables[$backendTable])) {
             self::$_attributeBackendTables[$backendTable] = $this->_getReadAdapter()->describeTable($backendTable);
         }
+
         $describe = self::$_attributeBackendTables[$backendTable];
         return $this->_getReadAdapter()->prepareColumnValue($describe['value'], $value);
     }
@@ -1534,6 +1547,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
         foreach ($row as $field => $value) {
             $whereArr[] = $adapter->quoteInto($field . '=?', $value);
         }
+
         $where = implode(' AND ', $whereArr);
 
         $adapter->beginTransaction();
@@ -1551,6 +1565,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
             } elseif ($origValueId !== false && ($newValue === null)) {
                 $adapter->delete($table, $where);
             }
+
             $this->_processAttributeValues();
             $adapter->commit();
         } catch (Exception $e) {
@@ -1578,18 +1593,20 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
 
         $this->_beforeDelete($object);
 
-        try {
-            $where = [
-                $this->getEntityIdField() . '=?' => $id,
-            ];
-            $this->_getWriteAdapter()->delete($this->getEntityTable(), $where);
-            $this->loadAllAttributes($object);
-            foreach (array_keys($this->getAttributesByTable()) as $table) {
-                // phpcs:ignore Ecg.Performance.Loop.ModelLSD
-                $this->_getWriteAdapter()->delete($table, $where);
+        if (isset($id)) {
+            try {
+                $where = [
+                    $this->getEntityIdField() . '=?' => $id,
+                ];
+                $this->_getWriteAdapter()->delete($this->getEntityTable(), $where);
+                $this->loadAllAttributes($object);
+                foreach (array_keys($this->getAttributesByTable()) as $table) {
+                    // phpcs:ignore Ecg.Performance.Loop.ModelLSD
+                    $this->_getWriteAdapter()->delete($table, $where);
+                }
+            } catch (Exception $e) {
+                throw $e;
             }
-        } catch (Exception $e) {
-            throw $e;
         }
 
         $this->_afterDelete($object);
